requirements.txt
psycopg2-binary==2.8.5
boto3==1.14.48
botocore==1.17.48
pytz==2020.1
